fix(robot-server): Better errors for bad wpa2 pass (#17203)
We pipe errors from nmcli straight to the display when we get them.
NMCLI has famously awful errors. We don't want to always squash them,
because they could be useful, but in the case of a mis-entered WPA2 PSK,
we can be pretty sure that's the problem.

This will now say "check your wifi credentials" 
<img width="607" alt="Screenshot 2025-01-07 at 4 46 15 PM"
src="https://github.com/user-attachments/assets/d9006f01-37e8-4221-812a-900ffcb61960"
/>


## Testing
- [x] Connect to a wifi network and purposefully enter the wrong
password; you should get a less awful message. Note that this only
actually works if you're disconnected from wifi when you try to connect;
if you're connected to wifi and then try and connect to a network and
use the wrong password, you just don't get a result because the request
went out over the original wifi network which got disconnected.

It's the same code on the flex, ot-2 on the desktop (the ODD has a
different message).
Closes RSQ-3