feat(shared-data): add deck schema v6
This schema contains a concept of "locating features" for addressable
areas. These locating features are things that constrain the position of
labware loaded on top of the addressable area. The presence of a
locating feature indicates that the addressable area can contain deck
items. Which locating feature is available constrains which labware can
be loaded; different labware define different locatable features (for
instance, their back left or front left corners, or the outer surface of
one of their wells).

Locating feature names are not constrained to enums because we expect a
fairly wide number of them.

Closes EXEC-80