fix(robot-server): Fix analysis log message (#17253)
robot-server logs a message when it begins a protocol analysis.
Historically, this happened when, and only when, a client hit the `POST
/protocols` endpoint, so the log was placed there. In modern times, it's
gotten a little more complicated, and the log has not caught up with
that: I think we are logging analyses that do not actually exist. This
fixes that.