fix(protocol-designer): fix save behavior for bad liquid form (#17262)
In liquids toolbox, we should never disable "save" when adding a liquid
to wells. Rather, errors should show when save is clicked, surfacing to
the user fields that need to be resolved. Here, I modify our logic for
Controller rules in our form used to added liquids, and implement a
volume validation function.

Closes RQA-3497