fix(api): Fix Pydantic error when parsing commands that did not succeed (#17264)
robot-server sometimes stores, in its database, run commands that do not
have a `result`. For example, if a command failed, it has an `error`
instead of a `result`.

After the recent Pydantic v1->v2 upgrade, parsing these commands was
broken for about 1/3rd of our command types. The server would raise a
500 error. Their Pydantic models were defined like:

```
result: Optional[FooResult]
```

In Pydantic v1, that parses JSON where `result` is `null` or omitted,
but in Pydantic v2, it only parses JSON where `result` is `null`. We
need to change it to:

```
result: Optional[FooResult] = None
```