feat(protocol-designer): initial absorbance reader stepform UI (#17280)
This PR adds the initial UI for absorbance reader steps. It creates the
new AbsorbanceReaderTools component in line with our other step form
toolbox components, and modularizes its children components for
cleanliness. The content of both pages 1 and 2 of the form are
determined by:
1) module initialization state, and
2) labware presence or lack thereof.

According to designs, the UI will funnel the user into the correct step
form type (lid, initializaton, or reading) based on those 2 inputs. Also
implicated here are creation and wiring up of the following methods:
- patch to pre-select an absorbance reader module if only one is
available
- form change handlers for absorbance reader dependent fields
- absorbance reader module option getter

Closes AUTH-1264