fix(protocol-designer): propagate mix form's name/description to args (#17275)
# Overview

We were not filling in the `name` and `description` in the `MixStepArgs`
with the data from the form, putting the dummy value `'description would
be here 2018-03-01'` there instead. This was apparently a TODO left over
from 2018.

Python generation is going to work off of the StepArgs, so I want all
the data I need to be in the StepArgs.

## Test Plan and Hands on Testing

I added a unit test.

## Review requests

Could you look at the other `MixStepArgs` values in the unit test and
see if they make sense? The other values are outside the scope of this
PR, but I'm not sure where values like `touchTipMmFromBottom: 9.54` are
coming from.

## Risk assessment

Should be low. I assume no one was using the `name` and `description`
before, because they would notice the blatant placeholder values if they
were.