feat(components, app): move CommandText and TimelineScrubber components (#17284)
This PR is part of very preliminary work to explore adding the timeline
scrubber to other projects other than only the app. In order to start
that work, the timeline scrubber & command text needed to be accessible
outside of the app so I moved those components and affected components
to the components directory.

Additionally, I added `step-generation` as a module for `components` and
I added `i18next` as a dependency, since command text has translations.