refactor(step-generation, protocol-designer): create HydratedFormData & update atomic command args to match v10 params (#17308)
Was cleaning up `StepArgs` a bit so it matches the command params a bit
better and then ended up creating `HydratedFormData` type and wired it
up. This resulted in a bit of a refactor and changing some arg names.

Additionally, this PR updates some of the atomic command types to using
the command's `Param` types to the latest command schema and updates the
stepArgs accordingly. The reason why I did this is because some atomic
command args were still using outdated params (from command schema V4,
V6, etc. When we are currently on command schema V10)