feat(protocol-designer): design cleanups and copy updates (#17318)
This PR makes minor fixes for the following issues:
1) add reference wavelength tooltip
2) remove instead of disabling button for adding wavelength when max are
added, and
3) using first known wavelength option rather than custom option when
adding new wavelength/reference wavelength.

This entails modifying the default form value for `wavelengths` field to
the first known wavelength (from a shared constant). I also add a patch
to update the reference wavelength to the first known wavelength when
the reference wavelength field is active.

Closes AUTH-1333