feat(protocol-designer, step-generation): add gripper error creator for absorbance reader lid commands (#17321)
This PR introduces a new error creator in step generation and wires up
to absorbance reader open and close lid commands. These commands should
now raise a gripper missing error if attempting to open or close the
absorbance reader lid without a gripper configured. Although presence of
a gripper is required to add an absorbance reader module or steps, an
error scenario can arise if a user correctly configures a gripper and
absorbance reader, and then removes the gripper from their protocol.

Closes AUTH-1194