feat(protocol-designer): handle (no) gripper with absorbance reader
This PR updates logic for adding absorbance reader module given presence or lack thereof of a gripper. During onboarding, if no gripper is added, the selector button for adding an absorbance reader module is disabled with tooltip. However, in deck setup, we no longer prevent the user from adding an absorbance reader module without the presence of a gripper. We will be permissive in setup, and present timeline errors in protocol timeline.

Closes AUTH-1345