fix(everything): replace Greek 'μ' with micro 'µ' (#17331)
# Overview

There are two different characters that look like the "mu" symbol:
- `µ` (U+00B5 `MICRO SIGN`)
- `μ` (U+03BC `GREEK SMALL LETTER MU`)

In some fonts they look different; but in many fonts they look the same.

The problem is that in the Public Sans font that we use for Protocol
Designer, `U+03BC GREEK SMALL LETTER MU` is not available at all,
because the font only supports Latin characters plus some scientific
symbols. So whenever we use `U+03BC GREEK SMALL LETTER MU` in text, the
browser has to render it with a fallback font, which can look ugly.

For consistency, we should just use `U+00B5 MICRO SIGN` whenever we have
a `µ` that means 1/1,000,000, and only use `U+03BC GREEK SMALL LETTER
MU` when we're writing Greek text (like in a Greek-language user
manual).

## Test Plan and Hands on Testing

I did a global search and replace on all the text in our codebase. I'm
relying on the CI tests to catch any issues.

## Risk assessment

This could potentially break external dependencies that expect the
Greek-text mu.