feat(protocol-designer): fix whitescreen and add preselect form type (#17335)
Here, I fix a whitescreen stemming from not correctly using optional
chaining when defining initialization. Also, I add a module required
error and wire up to the absorbance reader step form. Last, I fix the
preselect form type logic by adding a patch to createPresavedStepForm,
and also updating the useEffect logic in the step form component to not
fire on mount. We do this so that the form type is not updated if
opening a saved step form.

Closes AUTH-1343