refactor(api): location sequences for offsets
When we load and express labware offsets, we previously used an object
that had explicit named members for a labware parent; a module parent;
and a deck slot. This worked as long as we don't have heterogenous
labware stacking (to include labware-on-labware-on-adapter) but it's
inflexible, and maybe we do want that.

Instead, let's have a sequence of little models that specify a parent
geometry - a labware def uri, a module model, or an addressable area
name - and use that both for specifying offset locations via the HTTP
API, for comparing locations internally, and for downloading locations
from the HTTP API.

Importantly, these offset locations are currently in the database as
parts of runs, so we have to keep the old version around when expressing
labware data; this is annoying, but that's life. We can calculate the
equivalent old data when we load a labware offset pretty easily, and
this also lets us not change the legacy protocol core.

We also are going to keep the old method of specifying them around, and
convert them into the new format, which is also pretty easy, to preserve
backwards compatibility and roundtripping older offsets from before this
data was present.