refactor(robot-server): accept new offsets
Accept either old or new style offsets and adapt the internal APIs to
provide the correct type of thing to the protocol engine internals.

Note that the new data will essentially be lost when the data goes to
the labware offsets API because that API does not currently support
specifying (or returning) the location vector; that's fine in this PR
because all the APIs still accept the old kind of offset. In a follow-up
PR, we'll add support for the new-style locations to the offset api.