fix(protocol-designer): fix no tiprack-related whitescreen
This PR fixes a PD whitescreen bug where a user tries to create a transfer step wihout a tiprack available on the deck. As a temporary fix, I null check the first tiprack option in `TiprackField` and default to a placeholder string. This prevents a whitescreen, and creates a not-enough-tips timeline error if the step is saved, instructing the user to add tipracks to the deck to fix.

Closes RESC-380