refactor(api): Express labware offset locations as sequences (#17363)
This is a pattern that we're going to adopt for symbolic geometry
locations in HTTP API contexts - expressing them as a sequence of
heterogenous unions that represent kinds of places things can be. We'll
do this for locations of labware instances in a followup, but for now
we're doing it for offset locations.

Offset locations were previously objects that had a slot name; an
optional module model, for a location on a module; and an optional
labware URI, for a location on a labware (possibly an adapter, possibly
on a model). This can't represent labware stacks, and it's locked to
slot names. These are more or less fine for now still, but it would be
nice to support stacks in the general case, and it will be awkward for
clients to use the object-with-optional-attributes model alongside the
sequence-of-locations instance model. Plus, while this won't need a
database schema change for the offsets stored in runs, it _will_ for the
offsets api - so let's make this change now while that schema and API
haven't yet shipped.

## Testing
- [x] do the offsets still apply, on
   - [x] a slot
   - [x] a module
   - [x] an adapter on a model

Since the HTTP API still is the same and everything is tested
internally, it is sufficient to do normal LPC testing - everything
should still work.

## Reviews
- [ ] look appropriate? Missing anything?