fix(api): tune store_labware motion parameters and current to increase the z torque. (#17382)
Motion parameters need to be adjusted to account for tiprack weight by decreasing the speed when moving on the z-axis when storing labware. This pull request decreases the z speed to half the move speed and also decreases the current back down to 1.5mAmps since the lower speeds increase the torque.