refactor(step-generation, protocol-designer, app): touchTip to emit from top of well origin (#17375)
closes AUTH-1378

Previously, `touchTip` would occur from the bottom of the well, which is
a bit confusing since users in python and intuitively would want to
start with origin from the top. This PR migrates all touch tip fields
and inverses the logic. Here is the summary of the changes:
1. `aspirate_touchTipMmFromBottom` is now `aspirate_touchTipMmFromTop`
in `moveLiquid` form data
2. `dispense_touchTipMmFromBottom` is now `dispense_touchTipMmFromTop`
in `moveLiquid` form data
3. `mix_touchTipMmFromBottom` is now `mix_touchTipMmFromTop` in `mix`
form data
4. the modal in the form is updated to grab the values starting from the
top of the well, going into the well the values are negative. This
updated in PD **AND** Quick transfer
5. step-generation now emits `touchTip` to be origin from the top of the
well with the new z-index
6. old protocols now go through a `8_5_0` migration to migrate their
form data z-index fields to be origin from the top of the well