fix(api): use smaller of max pipette volume and max tip volume for splitting large transfer volumes (#17387)
# Overview

Found a bug in the logic that splits large volume transfers into smaller
volumes.
Bug is that we only check the transfer volume against max pipette
volume. So if you use a 50uL pipette with a 200uL tip, everything works
correctly, but if you use a 1000uL pipette with 50uL tip, it attempts to
pipette upto 1000uL, instead of clipping to the max tip volume of 50uL.

This PR fixes that by taking the smaller of pipette volume and tip
volume as the max volume a single transfer can take.

## Risk assessment

Low. A scope-limited bug fix