feat(protocol-designer): add python field to commands and timeline (#17383)
# Overview

This starts the plumbing for Python generation from Protocol Designer.
AUTH-1385

We're adding a field for Python code to each command
(`CommandsAndWarnings`) and to each entry of the timeline
(`CommandsAndRobotState`). And in the reducer, we concatenate the Python
commands together, analogously to how we concatenate the JSON commands
together.

## Test Plan and Hands on Testing

I added unit tests to show that the Python code concatenation works, and
that the Python code gets copied from the CommandCreators to the
Timeline. The tests also demonstrate that the changes do NOT affect the
behavior of existing commands that don't generate Python.

I've also done hands-on testing in a private experimental branch that
has a more complete implementation of Python generation.

## Review requests

My first time making a functional change to PD, let me know if I'm
overlooking anything.

## Risk assessment

Low. Should not cause any observable change to PD's behavior.