fix(step-generation): properly update robotState when dispense is after airGap (#17392)
closes RQA-3913

When `airGapinplace` was implemented, i incorrectly got rid of the
`dispense` `meta: isAirGap` information. This resulted in the
`robotState` liquid state incorrectly getting updated for dispense after
`airGapInPlace`. This PR adds the dispense `meta: isAirGap` back, purely
for the robot state purposes. Otherwise, i was not sure how to not
update the dispense robot state.