fix(protocol-designer): fix magnetic module errors
This PR properly displays error messages for bad engage heights with both the GEN1 and GEN2 magnetic modules in PD. The footprint here is large, because we need to pass the module model to the error creators that determine the module's valid engage height range (specific to model). We also need to pass moduleEntities down through createPresavedStepForm and its patches to set the module model when the form is created or the moduleId is updated.

Closes RQA-3911