feat(robot-server): handle new offset locations in /labwareOffsets (#17388)
Adds the new format for labware offset locations from #17363 to the HTTP
API `/labwareOffsets` endpoint.

The endpoint now accepts and returns labware offset locations as these
sequences. The filter API is for now unchanged, and still works properly
to filter out or in offsets based on their core features or on their
locations.

The types are defined separately from the protocol engine because we'll
be shortly adding a concept of a "general offset" to this API, and that
is best modeled as another heterogenous union entry, which means the
types will start to diverge.

This comes along with a sql schema migration; the offset locations are
normalized out into an offset table with a foreign key into the main
offset table. All content from the offset table is migrated to the new
one in the standard persistence directory migration api.

## Reviews
- does the sql look good?
- do the tests look appropriate? some have expanded
- do the migrations look good?

## Testing
- [x] put this on a machine that has offset data and make sure it
migrates and comes up with the data preserved

Closes EXEC-1105