fix(shared-data): tc lids are schema v2 now
We need to not use any schema v3 labware in 8.3 because we're going to
change the schema. The thermocycler lids were only available in schema
3. Change that by making the lid v1 def a schema 2 def, and making a lid
v2 def that is a schema 3 def.

All the data is the same, and we even kept the stackLimit and added it
to labware schema 2, so all the machine behavior should be exactly the
same. The only other thing that's not in schema 2 now is the new lid
stuff support, but that code isn't in 8.3 so we should be good.

Closes EXEC-1180