fix(protocol-designer): fix temperature step form
This PR fixes a bug in the temperature step form where a form error is incorrecly returned if a user has not entered a temperature, even if the setTemperature ield is set to false. The error arose from incorrectly parsing the string literals for 'true' / 'false'. Also, this PR fixes the step summary text for deactivating a temperature module

Closes RQA-3947, RQA-3948