docs(robot-server): Include `/labwareOffsets` endpoints in docs again (#17441)
These HTTP API endpoints were hidden from the docs because they were
incomplete and we didn't want the soon-to-be-released v8.3.0 software to
advertise them.

v8.3.0 is in its own branch now, so we can undo that, so the docs match
the in-development state of v8.4.0.

Reverts commit 5f869ea04c29cb6d1b2e4884af8d859dbbd88149, PR #17219.