fix(shared-data): evotips v1 are schema 2 now (#17428)
We had the evotips labware and adapter as schema 3 labware. This was
probably necessary while this was a 3-element stack, but now it is not,
and we need to avoid shipping schema 3 labware in 8.3 so that we retain
the ability to change it.

Closes EXEC-1179

## review
- seem fine?

## testing
- smoke test. nothing checks schema identity, just whether data is
there, and all the data that the code in 8.3 uses is still there.