fix(app): translate continue button in network (#17528)
The continue button "continue" text in various network settings panels
needs to come from the shared i18n source, which is where it is, not the
device_settings i18n source, which is where it isn't. By knowing where
continue isn't, we know where continue is, and by subtracting where
continue isn't from where continue is we know where continue should be,
and by making where continue should be go from where continue isn't to
where continue is, eventually we end up where continue is.

## testing
- put this odd on a robot
- from the settings screen, connect to wifi and check that continue is
localized. This should now happen in the "join other network" screen and
in the "select authentication type" screen
- from the first-run experience, try to connect via ethernet while an
ethernet cable is unplugged and note that "no data" is now localized
- from the first-run experience, try to connect via wifi and note that
the same things are now localized as in the setup screen

Closes RQA-3960