feat(api): use InStackerHopper for labware in...
stacker hoppers. Kept you guessing.

You don't specify this on the input to loadLabware because very soon you
won't be able to load labware into a stacker hopper anymore; once the
rest of the commands are implemented, labware will only get this
location if it is passed to stacker/store.

Closes EXEC-1209