feat(protocol-designer): bring back deck setup drag & drop (#17477)
closes AUTH-1276

This PR brings back drag/drop to deck setup with the same functionality
as it had in the pre-redesign. You can only drag top-level labware and
can not drag: modules or adapters. When we add stacking capabilities, we
will need to refactor it to account for that level of stacking.