feat(protocol-designer): Python generation for the Comment step (#17536)
# Overview

This adds Python generation for the Comment step in PD. The Comment step
has not been publicly released yet, but this was easy to implement, so I
might as well get it out of the way.

## Test Plan and Hands on Testing

Added unit tests (for the JSON implementation as well, which didn't seem
to have tests before). Confirmed the output when running from my branch.
Also checked that the generated Python file passes simulation.

## Risk assessment

Very low. Both comments and Python generation are hidden behind feature
flags.