fix(protocol-designer): don't emit define_liquid(description="") if it's empty string (#17538)
# Overview

The liquid description in the `LiquidEntities` seems to be the empty
string sometimes. We don't want to emit that into the Python code if
it's empty.

Before:
```
# Define Liquids:
liquid_1 = protocol.define_liquid(
    "Ketchup",
    description="",
    display_color="#ff4f4fff",
)
```

After:
```
# Define Liquids:
liquid_1 = protocol.define_liquid(
    "Ketchup",
    display_color="#ff4f4fff",
)
```
AUTH-1092

## Test Plan and Hands on Testing

Updated unit test to test the empty-string case.

## Risk assessment

Low. No one is relying on the generated Python code yet.