feat(api): add and use inStackerHopperLocation (#17535)
Labware objects that are in the stacker hopper have the new location
InStackerHopperLocation, which also identifies the labware Id of the
stacker in question.

In this PR, we set the location of labware that get loaded onto the
stacker module while static=False to InStackerHopperLocation. However,
after the rest of the work in the epic, the only way to get a labware
into `InStackerHopperLocation` is to pass it to `stacker/store`. It's
not a location you can specify in any parameters, which is already true
in this PR.

Because we're adding it fresh and thus it isn't already serialized
anywhere, we can give the location object a `kind` member and therefore
don't have to have a separate location sequence component.

## review requests
does this look like the right data? are we happy with it?

## testing
none required, this is well-covered by internal tests and the data that
it exposes that changed isn't actually used by anything yet.