feat(api): add setStoredLabware command (#17540)
Adds the new setStoredLabware command to constrain the flex stacker
labware pool.

The idea here is that you configure each stacker with some concept of
"this is the kind of labware you contain", and an amount of labware.
Then, you update how much it contains and when you retrieve stuff
`retrieve` creates a new labware object. Only labware that entered the
stacker through `store` retains its individuality.

This PR creates the command and adds the new data to the stacker
substate, but it's not yet used for anything; static mode remains,
loading labware objects into the hopper remains. In followup prs, we'll
remove that older loading capability, and switch things over to using
the new way of doing things.

## testing
- automated tests, really; the command does nothing

## reviews
- structure and arguments and such

Closes EXEC-1212

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: sfoster1 <3091648+sfoster1@users.noreply.github.com>