fix(protocol-designer): prevent liquid handling steps if no compatible labware
This PR refines `AddStepButton`'s supported steps logic for mix and moveLiquid, by checking the presence of compatible labware during the creation of the step. These two liquid handling options should only be shown if 1) the labware is on deck, 2) the labware is not an adapter, and 3) the labware is not a tiprack.

Closes RQA-3409